backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
aiosmtplib==3.0.1
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.25.2
slowapi==0.1.9

